requirements.txt
fastapi
uvicorn[standard]
openai
tweepy
sse-starlette
aiosqlite
jinja2
python-dotenv
